feat: remove src move to peak and then use poetry
