+++ pyproject.toml
@@ -1,16 +1,44 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-
-[project]
+[tool.poetry]
 name = "peak"
 version = "0.1.0"
-description = "A tool to calculate peak and non-peak hours of production"
-authors = [
-    { name="johnson2427", email="blakeejohnson39@gmail.com" }
-]
+description = "A time calculator"
+authors = ["johnson2427 <blakeejohnson39@gmail.com>"]
+license = "Apache 2.0"
 readme = "README.md"
-requires-python = ">=3.9"
-license = { text="Apache2.0" }
-dependencies = [
-    "pytest"
-]
+
+[tool.poetry.scripts]
+peak-cli = "src._cli:app"
+
+[tool.poetry.dependencies]
+python = "^3.12"
+click = "^8.1.7"
+
+[tool.poetry.dev-dependencies]
+pytest = "^8.3.3"
+black = "^24.8.0"
+isort = "^5.13.2"
+mypy = "^1.11.2"
+pytest-cov = "^5.0.0"
+
+[tool.mypy]
+exclude = "build/"
+
+[tool.black]
+line-length = 79
+target-version = ["py312"]
+include = '\.pyi?$'
+
+[tool.isort]
+line_length = 79
+force_grid_wrap = 0
+include_trailing_comma = true
+multi_line_output = 3
+use_parentheses = true
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+pythonpath = "peak/"
